Dynamically render room buttons (sort of) and added create room button which does not have any functionality yet
