+++ src/index.js
@@ -15,7 +15,7 @@ But this hasn't happened yet!
 import 'phaser';
 import io from 'socket.io-client';
 import config from './config/config';
-import MainScene from './scenes/MainScene';
+import LobbyScene from './scenes/LobbyScene';
 import FgScene from './scenes/FgScene';
 
 class Game extends Phaser.Game {
@@ -27,11 +27,11 @@ class Game extends Phaser.Game {
     this.socket = io();
 
     // add all the scenes
-    this.scene.add('MainScene', MainScene);
+    this.scene.add('LobbyScene', LobbyScene);
     this.scene.add('FgScene', FgScene);
 
     // start the game with the mainscene
-    this.scene.start('MainScene', { socket: this.socket });
+    this.scene.start('LobbyScene', { socket: this.socket });
   }
 }
